+++ app/src/main/java/pl/szczodrzynski/edziennik/api/v2/vulcan/Vulcan.kt
@@ -61,7 +61,7 @@ class Vulcan(val app: App, val profile: Profile?, val loginStore: LoginStore, va
         }
     }
 
-    override fun getMessage(messageId: Int) {
+    override fun getMessage(messageId: Long) {
 
     }
